Exclude CheckOps from interpreter interface
