+++ BUILD.bazel
@@ -402,6 +402,7 @@ cc_library(
     ],
     strip_include_prefix = ".",
     deps = [
+        ":interpreter_ops",
         ":reference_errors",
         ":reference_ops",
         ":reference_process",
@@ -422,7 +423,6 @@ cc_library(
     ],
     strip_include_prefix = ".",
     deps = [
-        ":interpreter_ops",
         ":reference_errors",
         ":reference_tensor",
         ":reference_token",
