+++ stablehlo/reference/Ops.cpp
@@ -17,7 +17,9 @@ limitations under the License.
 
 #include <algorithm>
 
+#include "llvm/ADT/APFloat.h"
 #include "llvm/ADT/APInt.h"
+#include "llvm/Support/Errc.h"
 #include "llvm/Support/Error.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/IR/BuiltinTypes.h"
