+++ stablehlo/tests/BUILD.bazel
@@ -34,7 +34,6 @@ cc_library(
         "//:base",
         "//:reference_interpretervalue",
         "//:reference_numpy",
-        "//:reference_scope",
         "//:reference_tensor",
         "//:reference_token",
     ],
